pyproject.toml
[project]
name = "media-service"
description = "Media service for bamper.by"
readme = "README.md"
version = "0.1.0"
requires-python = ">=3.10,<4"
dependencies = [
    "django>=5.2.6",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "django-unfold>=0.67.0",
    "psycopg[c]>=3.2.9",
    "requests>=2.32.5",
    "markdown>=3.9",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "django-stubs[compatible-mypy]>=5.2.2",
    "types-requests",
    "types-Markdown",
]

[tool.black]
target-version = ["py310", "py311", "py312", "py313"]
line-length = 120

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
exclude = ["venv", ".venv"]
disable_error_code = []

[[tool.mypy.overrides]]
module = [
    "rest_framework.*",
    "environ",
    "unfold.*",
    "faker.*",
]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "media_service.settings"
